+++ src/main/java/com/ani/taku_backend/marketprice/controller/CompletedDealController.java
@@ -43,17 +43,11 @@ public class CompletedDealController {
             @DateTimeFormat(pattern = "yyyy-MM-dd")
             @RequestParam LocalDate endDate,
 
-
             @Parameter(description = "그래프 표시 옵션 (기본값: ALL)", required = false)
             @RequestParam(defaultValue = "ALL", name = "displayOption") GraphDisplayOption displayOption,
 
-
-            @Parameter(description = "정렬 방향", schema = @Schema(allowableValues = {"ASC", "DESC"}))
-            @RequestParam(defaultValue = "ASC") Sort.Direction direction,
-
             @ParameterObject 
-            @PageableDefault(size = 5) Pageable pageable
-
+            @PageableDefault(size = 5, sort = "id", direction = Sort.Direction.ASC) Pageable pageable
     ) {
         try {
             var requestDTO = new PriceGraphRequestDTO(keyword, startDate, endDate, displayOption);
