+++ src/main/java/com/ani/taku_backend/post/service/PostService.java
@@ -1,6 +1,7 @@
 package com.ani.taku_backend.post.service;
 
 import com.ani.taku_backend.post.model.dto.PostCreateRequestDTO;
+import com.ani.taku_backend.post.model.dto.PostDetailResponseDTO;
 import com.ani.taku_backend.post.model.dto.PostListRequestDTO;
 import com.ani.taku_backend.post.model.dto.PostListResponseDTO;
 import com.ani.taku_backend.post.model.dto.PostUpdateRequestDTO;
@@ -15,4 +16,6 @@ public interface PostService {
     Long updatePost(Long postId, PostUpdateRequestDTO postUpdateRequestDTO, User user);
 
     void deletePost(Long postId, User user);
+
+    PostDetailResponseDTO getPostDetail(Long postId, boolean canAddView, Long currentUserId);
 }
